spirolight function
