+++ xmaslights-spin.py
@@ -1,11 +1,8 @@
-# Here are the libraries I am currently using:
-import time
-# import board
-# import neopixel  # pip install rpi_ws281x adafruit-circuitpython-neopixel adafruit-blinka
-import math
-from collections import namedtuple  # for Point data structure
-import pygame  # for virtual tree visuals
+# refactored by pikuch 24.12.2020
 
+import time
+import math
+from xmastrees import VirtualXmasTree, XmasTree
 
 # You are welcome to add any of these:
 # import random
@@ -18,83 +15,7 @@ import pygame  # for virtual tree visuals
 # some_value = int(sys.argv[0])
 
 
-# class XmasTree:
-#     def __init__(self, coord_filename):
-#         self.coords = self.read_coordinates(coord_filename)
-#         self.min_x = min(coord.x for coord in self.coords)
-#         self.max_x = max(coord.x for coord in self.coords)
-#         self.min_y = min(coord.y for coord in self.coords)
-#         self.max_y = max(coord.y for coord in self.coords)
-#         self.min_z = min(coord.z for coord in self.coords)
-#         self.max_z = max(coord.z for coord in self.coords)
-#
-#         self.PIXEL_COUNT = len(self.coords)
-#         assert (self.PIXEL_COUNT == 500)
-#         self.pixels = neopixel.NeoPixel(board.D18, self.PIXEL_COUNT, auto_write=False)
-#
-#     def read_coordinates(self, coord_filename):
-#         Point = namedtuple("Point", ["x", "y", "z"])  # so we can address point coordinates like humans
-#         with open(coord_filename, 'r') as fin:
-#             coords_raw = fin.readlines()
-#         return list(map(lambda line: Point(*map(lambda item: int(item.strip()), line[1:-2].split(","))), coords_raw))
-#
-#     def display(self):
-#         self.pixels.show()
-#
-#     def set_led_RGB(self, n, RGBcolor):
-#         self.pixels[n] = [RGBcolor[1], RGBcolor[0], RGBcolor[2]]
-
-
-class VirtualXmasTree:
-    def __init__(self, coord_filename):
-        self.coords = self.read_coordinates(coord_filename)
-        self.min_x = min(coord.x for coord in self.coords)
-        self.max_x = max(coord.x for coord in self.coords)
-        self.min_y = min(coord.y for coord in self.coords)
-        self.max_y = max(coord.y for coord in self.coords)
-        self.min_z = min(coord.z for coord in self.coords)
-        self.max_z = max(coord.z for coord in self.coords)
-
-        self.PIXEL_COUNT = len(self.coords)
-        assert (self.PIXEL_COUNT == 500)
-        self.pixels = [[0, 0, 0]] * self.PIXEL_COUNT
-
-        self.tree_size = self.max_z - self.min_z
-        self.max_light_size = 10
-        self.min_light_size = 4
-        self.margin = self.tree_size * 0.1
-        self.WINDOW_WIDTH, self.WINDOW_HEIGHT = 800, 800
-        pygame.init()
-        self.clock = pygame.time.Clock()
-        self.screen = pygame.display.set_mode((self.WINDOW_WIDTH, self.WINDOW_HEIGHT))
-        pygame.display.set_caption("Virtual Xmas Tree")
-
-    def read_coordinates(self, coord_filename):
-        Point = namedtuple("Point", ["x", "y", "z"])  # so we can address point coordinates like humans
-        with open(coord_filename, 'r') as fin:
-            coords_raw = fin.readlines()
-        return list(map(lambda line: Point(*map(lambda item: int(item.strip()), line[1:-2].split(","))), coords_raw))
-
-    def display(self):
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                exit(0)
-
-        self.screen.fill((127, 127, 127))
-
-        for p in range(len(self.coords)):
-            screen_x = int(self.margin + self.WINDOW_WIDTH * ((self.coords[p].x - self.min_x) / self.tree_size))
-            screen_y = int(self.WINDOW_HEIGHT - self.margin - self.WINDOW_HEIGHT * ((self.coords[p].z - self.min_z) / self.tree_size))
-            on_screen_size = int(self.min_light_size + (self.coords[p].y - self.min_y) / (self.max_y - self.min_y) * (self.max_light_size - self.min_light_size))
-            pygame.draw.circle(self.screen, self.pixels[p], (screen_x, screen_y), on_screen_size)
-
-        pygame.display.flip()
-        self.clock.tick(60)
-
-    def set_led_RGB(self, n, RGBcolor):
-        self.pixels[n] = RGBcolor
-
-
+# original function
 def xmaslight(tree):
 
     # VARIOUS SETTINGS
@@ -177,11 +98,35 @@ def xmaslight(tree):
             direction = -1
 
 
+def spirolight(tree):
+    # starting angle for each color channel (in radians)
+    angles = [0.0] * 3
+    # how much the angle changes per cycle
+    delta_angles = [1/5, 1/7, 1/9]
+    # maximum intensity of any one colour channel
+    max_intensity = 100
+
+    while True:
+        for led in range(len(tree.coords)):
+            rotation = math.atan2(tree.coords[led].y, tree.coords[led].x)
+            colour = [max_intensity * (math.sin(angles[0] + rotation) + 1) / 2,
+                      max_intensity * (math.sin(angles[1] + rotation) + 1) / 2,
+                      max_intensity * (math.sin(angles[2] + rotation) + 1) / 2]
+            tree.set_led_RGB(led, colour)
+
+        tree.display()
+
+        # update angles
+        for i in range(len(angles)):
+            angles[i] = math.fmod(angles[i] + delta_angles[i], 2 * math.pi)
+
+
 def run_lights():
     coord_filename = "Python/coords.txt"
-    # tree = XmasTree(coord_filename)
     tree = VirtualXmasTree(coord_filename)
-    xmaslight(tree)  # we pass the display to the function and let it run
+    spirolight(tree)
+    # tree = XmasTree(coord_filename)
+    # xmaslight(tree)
 
 
 if __name__ == "__main__":
